Added field generator

+++ src/DragonFly.API/Extensions/AssetApiExtensions.cs
@@ -7,6 +7,7 @@ using DragonFly.Permissions;
 using DragonFly.Query;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.Net.Http.Headers;
 
@@ -37,20 +38,32 @@ static class AssetApiExtensions
         return queryResult.Convert(x => x.ToRest());
     }
 
-    private static async Task<RestAsset> MapGet(HttpContext context, IAssetStorage storage, Guid id)
+    private static async Task<Results<Ok<RestAsset>, NotFound>> MapGet(HttpContext context, IAssetStorage storage, Guid id)
     {
-        Asset entity = await storage.GetRequiredAssetAsync(id);
+        Asset? entity = await storage.GetAssetAsync(id);
+
+        if (entity == null)
+        {
+            return TypedResults.NotFound();
+        }
 
         RestAsset restAsset = entity.ToRest();
 
-        return restAsset;
+        return TypedResults.Ok(restAsset);
     }
 
-    private static async Task MapDelete(HttpContext context, IAssetStorage storage, Guid id)
+    private static async Task<Results<Ok, NotFound>> MapDelete(HttpContext context, IAssetStorage storage, Guid id)
     {
-        Asset entity = await storage.GetRequiredAssetAsync(id);
+        Asset? entity = await storage.GetAssetAsync(id);
+
+        if (entity == null)
+        {
+            return TypedResults.NotFound();
+        }
 
         await storage.DeleteAsync(entity);
+
+        return TypedResults.Ok();
     }
 
     private static async Task<ResourceCreated> MapCreate(HttpContext context, IAssetStorage storage, RestAsset restAsset)
@@ -69,48 +82,74 @@ static class AssetApiExtensions
         await storage.UpdateAsync(asset);
     }
 
-    private static async Task MapPublish(HttpContext context, IAssetStorage storage, Guid id)
+    private static async Task<Results<Ok, NotFound>> MapPublish(HttpContext context, IAssetStorage storage, Guid id)
     {
-        Asset entity = await storage.GetRequiredAssetAsync(id);
+        Asset? entity = await storage.GetAssetAsync(id);
+
+        if (entity == null)
+        {
+            return TypedResults.NotFound();
+        }
 
         await storage.PublishAsync(entity);
+
+        return TypedResults.Ok();
     }
 
-    private static async Task<IResult> MapDownload(HttpContext context, IAssetStorage storage, Guid id)
+    private static async Task<Results<FileStreamHttpResult, NotFound, StatusCodeHttpResult>> MapDownload(HttpContext context, IAssetStorage storage, Guid id)
     {
-        Asset asset = await storage.GetRequiredAssetAsync(id);
+        Asset? asset = await storage.GetAssetAsync(id);
+
+        if (asset == null)
+        {
+            return TypedResults.NotFound();
+        }
 
         EntityTagHeaderValue etag = new EntityTagHeaderValue($"\"{asset.Hash}\"");
 
         if (context.Request.Headers.ETag == etag)
         {
-            return Results.StatusCode(StatusCodes.Status304NotModified);
+            return TypedResults.StatusCode(StatusCodes.Status304NotModified);
         }
 
         context.Response.GetTypedHeaders().CacheControl = new CacheControlHeaderValue() { Public = true, MaxAge = TimeSpan.FromDays(30) };
 
         Stream assetStream = await storage.GetStreamAsync(asset);
 
-        return Results.Stream(assetStream, contentType : asset.MimeType, entityTag: etag);
+        return TypedResults.Stream(assetStream, contentType : asset.MimeType, entityTag: etag);
     }
 
-    private static async Task MapUpload(HttpContext context, IAssetStorage storage, Guid id)
+    private static async Task<Results<Ok, NotFound>> MapUpload(HttpContext context, IAssetStorage storage, Guid id)
     {
         if (context.Request.ContentType == null)
         {
             throw new Exception("Content-type was not set.");
         }
 
-        Asset asset = await storage.GetRequiredAssetAsync(id);
+        Asset? asset = await storage.GetAssetAsync(id);
+
+        if (asset == null)
+        {
+            return TypedResults.NotFound();
+        }
 
         await storage.UploadAsync(asset, context.Request.ContentType, context.Request.Body);
+
+        return TypedResults.Ok();
     }
 
-    private static async Task MapRefreshMetadata(HttpContext context, IAssetStorage storage, Guid id)
+    private static async Task<Results<Ok, NotFound>> MapRefreshMetadata(HttpContext context, IAssetStorage storage, Guid id)
     {
-        Asset asset = await storage.GetRequiredAssetAsync(id);
+        Asset? asset = await storage.GetAssetAsync(id);
+
+        if (asset == null)
+        {
+            return TypedResults.NotFound();
+        }
 
         await storage.ApplyMetadataAsync(asset);
+
+        return TypedResults.Ok();
     }
 
     private static async Task<IBackgroundTaskInfo> MapRefreshMetadataQuery(HttpContext context, IAssetStorage storage, AssetQuery query)
